+++ tests/tst/errors/InvalidConfig.php
@@ -3,7 +3,7 @@
  * @package axy\errors
  */
 
-namespace axy\errors\tests\nstst\errors;
+namespace axy\errors\tests\tst\errors;
 
 class InvalidConfig extends \axy\errors\InvalidConfig implements Error
 {