+++ tests/tst/errors/Pointless.php
@@ -3,7 +3,7 @@
  * @package axy\errors
  */
 
-namespace axy\errors\tests\nstst\errors;
+namespace axy\errors\tests\tst\errors;
 
 class Pointless extends \axy\errors\Pointless implements Error
 {